refactor: コンポーネントごとに専用ディレクトリを作成しAGENTS.mdにルールを明記
