+++ AGENTS.md
@@ -88,6 +88,7 @@ src/
 ```
 
 - コンポーネントの名前はPascalCaseで命名し、ディレクトリ名はkebab-caseで命名してください。
+- 一つのディレクトリに複数のコンポーネントを配置してはなりません。コンポーネントごとに `{component-name}/` ディレクトリを作成し、`index.ts` からエクスポートしてください。
 
 ### Feature 内モジュールの参照制限
