+++ src/features/task/components/task-detail-skeleton/index.ts
@@ -1,0 +1,1 @@
+export { TaskDetailSkeleton } from "./task-detail-skeleton";
