+++ src/features/task/components/task-list-fetcher/index.ts
@@ -1,0 +1,1 @@
+export { TaskListFetcher } from "./task-list-fetcher";
