+++ src/features/task/components/task-list-fetcher/task-list-fetcher.tsx
@@ -1,7 +1,7 @@
 import "server-only";
 import { getTasks } from "#/features/task/actions/get-tasks";
 
-import { TaskList } from ".";
+import { TaskList } from "../task-list";
 
 import type { TaskStatus } from "#/features/task/types/task";
