+++ src/features/task/components/task-list-skeleton/index.ts
@@ -1,0 +1,1 @@
+export { TaskListSkeleton } from "./task-list-skeleton";
