+++ src/features/task/components/task-tabs-fetcher/index.ts
@@ -1,0 +1,1 @@
+export { TaskTabsFetcher } from "./task-tabs-fetcher";
