+++ src/features/task/components/task-tabs/index.ts
@@ -1,0 +1,1 @@
+export { TaskTabs } from "./task-tabs";
